requirements.txt
fastapi
uvicorn
google-cloud-bigquery
ulid-py
google-cloud-storage